Miraculously - Refactored without breaking functionality - Moved initializations of stuff into initialization classes - Created 'ConfigInit' - Created 'ConsoleInit' - Created 'TwitchListeners' - Modified Polly setup and renamed to PollyInit - Each initializer holds it's respective related important objects (TwitchClient, PollyHandler, InputManager, The ConfigYmls)
