+++ src/main/java/simplexity/commands/HelpCommand.java
@@ -3,8 +3,8 @@ package simplexity.commands;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.event.Level;
-import simplexity.Main;
 import simplexity.config.LocaleHandler;
+import simplexity.console.ConsoleInit;
 import simplexity.console.Logging;
 
 public class HelpCommand extends Command {
@@ -17,7 +17,7 @@ public class HelpCommand extends Command {
     @Override
     public void execute() {
         Logging.logAndPrint(logger, LocaleHandler.getInstance().getHelpHeader(), Level.INFO);
-        for (Command command : Main.getCommandManager().getCommands().values()) {
+        for (Command command : ConsoleInit.getCommandManager().getCommands().values()) {
             String commandHelpMessage = LocaleHandler.getInstance().getHelpCommands().replace("%command_name%", command.getName())
                     .replace("%command_description%", command.getDescription());
             Logging.logAndPrint(logger, commandHelpMessage, Level.INFO);
